rename, move and import instance details and answer selector modals
