+++ app/screens/modals/instance-details-modal.tsx
